+++ Enderlook.Pools/src/Instance/ExactLengthArrayPoolAdapter.cs
@@ -1,11 +1,13 @@
 ﻿namespace Enderlook.Pools;
 
 /// <summary>
-/// Represent an adapter of <see cref="ExactLengthArrayPool{T}"/> to <see cref="ObjectPool{T}"/>.
+/// Represent an adapter of <see cref="ExactLengthArrayPool{T}"/> to <see cref="ArrayObjectPool{T}"/>.
 /// </summary>
-/// <typeparam name="TElement"></typeparam>
-internal sealed class ExactLengthArrayPoolAdapter<TElement>(ExactLengthArrayPool<TElement> pool, int length) : ObjectPool<TElement[]>
+internal sealed class ExactLengthArrayPoolAdapter<TElement>(ExactLengthArrayPool<TElement> pool, int length, bool clearOnReturn) : ArrayObjectPool<TElement>
 {
+    /// <inheritdoc cref="ArrayObjectPool{T}.ShouldClearArrayOnReturnByDefault"/>
+    public override bool ShouldClearArrayOnReturnByDefault => clearOnReturn;
+
     /// <inheritdoc cref="ObjectPool{T}.ApproximateCount"/>
     public override int ApproximateCount() => pool.ApproximateCount();
 
@@ -19,7 +21,17 @@ internal sealed class ExactLengthArrayPoolAdapter<TElement>(ExactLengthArrayPool
             Utils.ThrowArgumentOutOfRangeException_ArrayLength();
 
         // We allow passing `null`s so they are handled by the actual implementation.
-        pool.Return(element!);
+        pool.Return(element!, clearOnReturn);
+    }
+
+    /// <inheritdoc cref="ArrayObjectPool{T}.Return(T[], bool)"/>
+    public override void Return(TElement[] element, bool clearArray)
+    {
+        if (element?.Length != length)
+            Utils.ThrowArgumentOutOfRangeException_ArrayLength();
+
+        // We allow passing `null`s so they are handled by the actual implementation.
+        pool.Return(element!, clearArray);
     }
 
     /// <inheritdoc cref="ObjectPool{T}.Trim(bool)"/>